Run tests on Babel 6 and Babel 7
