+++ fixtures/assert-expansion/expectation6.js
