+++ fixtures/assert-expansion/expectation7.js
@@ -1,0 +1,2 @@
+(true && !((() => true)()) && console.assert((() => true)(), 'This is an assertion'));
+(true && !(false) && console.assert(false, 'This is an assertion 2'));
