+++ fixtures/debug-flag/expectation6.js
@@ -1,0 +1,5 @@
+
+
+if (true) {
+  console.log('woot');
+}
