+++ fixtures/debug-flag/expectation7.js
