+++ fixtures/default-export-features/expectation7.js
