+++ fixtures/deprecate-expansion/expectation6.js
