+++ fixtures/deprecate-expansion/expectation7.js
@@ -1,0 +1,2 @@
+(true && !(true) && console.warn('This is deprecated'));
+(true && !(false) && console.warn('Message without predicate'));
