+++ fixtures/development-feature-flags/expectation7.js
@@ -1,0 +1,8 @@
+import { FEATURE_A, FEATURE_B } from '@ember/features';
+
+let a;
+if (FEATURE_A) {
+  a = () => console.log('hello');
+} else if (FEATURE_B) {
+  a = () => console.log('bye');
+}
