+++ fixtures/development-svelte-builds/expectation7.js
