+++ fixtures/does-not-modify-non-imported-flags/expectation6.js
