+++ fixtures/does-not-modify-non-imported-flags/expectation7.js
@@ -1,0 +1,5 @@
+const DEBUG = 1;
+
+if (DEBUG) {
+  console.log('do it!');
+}
