+++ fixtures/ember-cli-babel-config/expectation7.js
