+++ fixtures/global-external-helpers/expectation6.js
