+++ fixtures/global-external-helpers/expectation7.js
@@ -1,0 +1,7 @@
+(true && __debugHelpers__.warn('This is a warning'));
+(true && __debugHelpers__.assert('Hahahaha', foo));
+(true && !(true) && __debugHelpers__.deprecate('This thing is donzo', true, {
+  id: 'donzo',
+  until: '4.0.0',
+  url: 'http://example.com'
+}));
