+++ fixtures/hygenic-debug-injection/expectation6.js
@@ -1,0 +1,7 @@
+
+import bar from 'something';
+
+if (bar()) {
+  const DEBUG = 'hahah';
+  (true && !(true) && console.warn('This is deprecated'));
+}
