+++ fixtures/hygenic-debug-injection/expectation7.js
