+++ fixtures/inject-env-flags/expectation6.js
@@ -1,0 +1,11 @@
+
+
+let testing;
+if (false) {
+  testing = 'WOOT';
+}
+
+let debug;
+if (true) {
+  debug = 'DEBUG';
+}
