+++ fixtures/inline-feature-flags/expectation6.js
@@ -1,0 +1,18 @@
+
+
+let a;
+if (false) {
+  a = () => console.log('hello');
+} else if (true) {
+  a = () => console.log('bye');
+}
+
+if (!false) {
+  console.log('stuff');
+}
+
+a = false ? 'hello' : 'bye';
+
+if (false && window.foo && window.bar) {
+  console.log('wheeee');
+}
