+++ fixtures/inline-feature-flags/expectation7.js
