+++ fixtures/log-expansion/expectation6.js
