+++ fixtures/log-expansion/expectation7.js
@@ -1,0 +1,1 @@
+(true && console.log('This is a logging'));
