+++ fixtures/production-svelte-builds/expectation6.js
@@ -1,0 +1,25 @@
+
+
+export let PartialComponentManager;
+if (false) {
+  PartialComponentManager = class {
+    constructor() {
+      this.isDone = true;
+    }
+  };
+}
+
+export let ObjectController;
+if (true) {
+  ObjectController = class {
+    constructor() {
+      this.isDoneAsWell = true;
+    }
+  };
+}
+
+export default class TheThingToReplaceControllers {
+  constructor() {
+    this.isNew = true;
+  }
+}
