+++ fixtures/production-svelte-builds/expectation7.js
