+++ fixtures/retain-module-external-helpers/expectation7.js
