+++ fixtures/retains-import-for-non-macro-types/expectation6.js
@@ -1,0 +1,4 @@
+
+import { isTesting } from '@ember/debug-tools';
+
+(true && console.warn('This is a warning'));
