+++ fixtures/retains-import-for-non-macro-types/expectation7.js
