+++ fixtures/retains-runtime-definitions/expectation6.js
@@ -1,0 +1,5 @@
+const DEPRECATIONS = { 'foo-bar': '1.2.0' };
+const FEATURES = { 'widget-woo': true };
+
+export const WIDGET_WOO = FEATURES['widget-woo'];
+export const FOO_BAR = DEPRECATIONS['foo-bar'];
