+++ fixtures/retains-runtime-definitions/expectation7.js
