+++ fixtures/runtime-feature-flags/expectation7.js
@@ -1,0 +1,9 @@
+import { FEATURE_B } from '@ember/features';
+
+if (true) {
+  console.log('woot');
+}
+
+if (FEATURE_B) {
+  console.log('aw yea');
+}
