+++ fixtures/shared-debug-module/expectation6.js
@@ -1,0 +1,5 @@
+var inspect = Ember.inspect;
+
+
+inspect('string thing');
+(true && !(false) && Ember.assert(false, 'This is an assertion 2'));
