+++ fixtures/shared-debug-module/expectation7.js
