+++ fixtures/warn-expansion/expectation6.js
