+++ fixtures/warn-expansion/expectation7.js
@@ -1,0 +1,1 @@
+(true && console.warn('This is a warning'));
