+++ package.json
@@ -2,7 +2,8 @@
   "devDependencies": {
     "@babel/core": "7.0.0-beta.42",
     "@babel/preset-env": "7.0.0-beta.42",
-    "@babel/register": "7.0.0-beta.42",
+    "babel-core": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
     "jest": "^21.0.0"
   },
   "name": "babel-plugin-debug-macros",
