+++ src/tests/debug-tools-babel7-test.js
@@ -1,0 +1,18 @@
+'use strict';
+
+const createTests = require('./create-tests');
+const transform = require('@babel/core').transform;
+
+const presets = [["@babel/preset-env", {
+  "targets": {
+    "browsers": ["> 5%"],
+  },
+  "modules": false
+}]];
+
+
+createTests({
+  presets: presets,
+  babelVersion: 7,
+  transform: transform
+});
